Last of the Big Commits (it's a terrible habit).
Munged the source tree into a structure I'm more comfortable with.

Added various pretty arts because it was more fun than real coding.

Added a couple of utility classes - Resource to allow easy loading of ipad versions of resources when the app is run on ipad, and Configuration, a singleton which will read and make available any plist based configs for the app. Which leads me to...

Made the initial scene loaded configuration driven. The scene class to use on launch is now specified in Main.plist, and then created and run.

Removed some of the standard fluff from the template.

Enabled universal and retina modes.